requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
aiohttp==3.9.1
langchain
langchain-community
faiss-cpu
openai
pypdf
unstructured
markdown
python-magic-bin
tiktoken
